requirements.txt
pytest
pytest-asyncio
pytest-mock
pyautogen==0.1.13
pyautogen[retrievechat]==0.1.13
matplotlib
scikit-learn
torch
tensorboard
streamlit
ipython

# llama tinygrad integration requirements
sentencepiece

# Used for Mac's Metal GPU support
# pyobjc-framework-Metal
# pyobjc-framework-Cocoa
# pyobjc-framework-libdispatch

# required for gpt2 model
tiktoken

# Required for whisper
# requires apt-get install portaudio19-dev || brew install portaudio
# WSL specific follow this guide to enable USB devices if your mic is not working
#   to fix it https://learn.microsoft.com/en-us/windows/wsl/connect-usb
librosa
pyaudio